+++ src/tests/solidityTests/emptyTestContract/test.emptyNotPayableTest.sol
@@ -1,0 +1,6 @@
+pragma solidity ^0.4.4;
+
+import "wafr/Test.sol";
+
+contract NoPayableTest {
+}